+++ src/lib.rs
@@ -3,6 +3,10 @@
 pub use embedded_nal;
 pub use smoltcp;
 
+use smoltcp::dhcp::Dhcpv4Client;
+use smoltcp::socket::AnySocket;
+use smoltcp::wire::{IpAddress, IpCidr, Ipv4Address, Ipv4Cidr};
+
 use core::cell::RefCell;
 use heapless::{consts, Vec};
 use nanorand::{wyrand::WyRand, RNG};
@@ -14,6 +18,7 @@ pub enum NetworkError {
     ReadFailure,
     WriteFailure,
     Unsupported,
+    NoIpAddress,
 }
 
 ///! Network abstraction layer for smoltcp.
@@ -22,10 +27,12 @@ where
     DeviceT: for<'c> smoltcp::phy::Device<'c>,
 {
     network_interface: RefCell<smoltcp::iface::EthernetInterface<'b, DeviceT>>,
+    dhcp_client: RefCell<Option<Dhcpv4Client>>,
     sockets: RefCell<smoltcp::socket::SocketSet<'a>>,
     used_ports: RefCell<Vec<u16, consts::U16>>,
     unused_handles: RefCell<Vec<smoltcp::socket::SocketHandle, consts::U16>>,
     randomizer: RefCell<WyRand>,
+    name_servers: RefCell<[Option<smoltcp::wire::Ipv4Address>; 3]>,
 }
 
 impl<'a, 'b, DeviceT> NetworkStack<'a, 'b, DeviceT>
@@ -34,27 +41,42 @@ where
 {
     /// Construct a new network stack.
     ///
+    /// # Note
+    /// This implementation only supports up to 16 usable sockets.
+    ///
+    /// Any handles provided to this function must not be used after constructing the network
+    /// stack.
+    ///
+    /// This implementation currently only supports IPv4.
+    ///
     /// # Args
     /// * `stack` - The ethernet interface to construct the network stack from.
     /// * `sockets` - The socket set to contain any socket state for the stack.
+    /// * `handles` - A list of socket handles that can be used.
+    /// * `dhcp` - An optional DHCP client if DHCP usage is desired. If None, DHCP will not be used.
     ///
     /// # Returns
     /// A embedded-nal-compatible network stack.
     pub fn new(
         stack: smoltcp::iface::EthernetInterface<'b, DeviceT>,
         sockets: smoltcp::socket::SocketSet<'a>,
+        handles: &[smoltcp::socket::SocketHandle],
+        dhcp: Option<Dhcpv4Client>,
     ) -> Self {
         let mut unused_handles: Vec<smoltcp::socket::SocketHandle, consts::U16> = Vec::new();
-        for socket in sockets.iter() {
-            unused_handles.push(socket.handle()).unwrap();
+        for handle in handles.iter() {
+            // Note: If the user supplies too many handles, we choose to silently drop them.
+            unused_handles.push(*handle).ok();
         }
 
         NetworkStack {
             network_interface: RefCell::new(stack),
             sockets: RefCell::new(sockets),
             used_ports: RefCell::new(Vec::new()),
-            unused_handles: RefCell::new(unused_handles),
             randomizer: RefCell::new(WyRand::new()),
+            dhcp_client: RefCell::new(dhcp),
+            unused_handles: RefCell::new(unused_handles),
+            name_servers: RefCell::new([None, None, None]),
         }
     }
 
@@ -70,18 +92,99 @@ where
     ///
     /// # Returns
     /// A boolean indicating if the network stack updated in any way.
-    pub fn poll(&self, time: u32) -> bool {
-        match self.network_interface.borrow_mut().poll(
-            &mut self.sockets.borrow_mut(),
-            smoltcp::time::Instant::from_millis(time as i64),
-        ) {
+    pub fn poll(&self, time: u32) -> Result<bool, smoltcp::Error> {
+        let now = smoltcp::time::Instant::from_millis(time as i64);
+        let updated = match self
+            .network_interface
+            .borrow_mut()
+            .poll(&mut self.sockets.borrow_mut(), now)
+        {
             Ok(updated) => updated,
+            err => return err,
+        };
 
-            // Ignore any errors - they indicate failures with external reception as opposed to
-            // internal state management.
-            // TODO: Should we ignore `Exhausted`, `Illegal`, `Unaddressable`?
-            Err(_) => true,
+        // Service the DHCP client.
+        if let Some(dhcp_client) = &mut *self.dhcp_client.borrow_mut() {
+            let mut interface = self.network_interface.borrow_mut();
+            let mut sockets = self.sockets.borrow_mut();
+            match dhcp_client.poll(&mut interface, &mut sockets, now) {
+                Ok(Some(config)) => {
+                    if let Some(cidr) = config.address {
+                        if cidr.address().is_unicast() {
+                            // Note(unwrap): This stack only supports IPv4 and the client must have
+                            // provided an address.
+                            if cidr.address().is_unspecified()
+                                || interface.ipv4_address().unwrap() != cidr.address()
+                            {
+                                // If our address has updated or is not specified, close all
+                                // sockets. Note that we have to ensure that the sockets we borrowed
+                                // earlier are now returned.
+                                drop(sockets);
+                                self.close_sockets();
+
+                                interface.update_ip_addrs(|addrs| {
+                                    // Note(unwrap): This stack requires at least 1 Ipv4 Address.
+                                    let addr = addrs
+                                        .iter_mut()
+                                        .filter(|cidr| match cidr.address() {
+                                            IpAddress::Ipv4(_) => true,
+                                            _ => false,
+                                        })
+                                        .next()
+                                        .unwrap();
+
+                                    *addr = IpCidr::Ipv4(cidr);
+                                });
+                            }
+                        }
+                    }
+
+                    // Store DNS server addresses for later read-back
+                    *self.name_servers.borrow_mut() = config.dns_servers;
+
+                    if let Some(route) = config.router {
+                        // Note: If the user did not provide enough route storage, we may not be
+                        // able to store the gateway.
+                        interface.routes_mut().add_default_ipv4_route(route)?;
+                    }
+                }
+                Ok(None) => {}
+                Err(err) => return Err(err),
+            }
         }
+
+        Ok(updated)
+    }
+
+    /// Force-close all sockets.
+    pub fn close_sockets(&self) {
+        // Close all sockets.
+        for mut socket in self.sockets.borrow_mut().iter_mut() {
+            // We only explicitly can close TCP sockets because we cannot access other socket types.
+            if let Some(ref mut socket) =
+                smoltcp::socket::TcpSocket::downcast(smoltcp::socket::SocketRef::new(&mut socket))
+            {
+                socket.abort();
+            }
+        }
+    }
+
+    /// Handle a disconnection of the physical interface.
+    pub fn handle_link_reset(&mut self) {
+        // Reset the DHCP client.
+        if let Some(ref mut client) = *self.dhcp_client.borrow_mut() {
+            client.reset(smoltcp::time::Instant::from_millis(-1));
+        }
+
+        // Close all of the sockets and de-configure the interface.
+        self.close_sockets();
+
+        let mut interface = self.network_interface.borrow_mut();
+        interface.update_ip_addrs(|addrs| {
+            addrs.iter_mut().next().map(|addr| {
+                *addr = IpCidr::Ipv4(Ipv4Cidr::new(Ipv4Address::UNSPECIFIED, 0));
+            });
+        });
     }
 
     // Get an ephemeral TCP port number.
@@ -106,6 +209,15 @@ where
             }
         }
     }
+
+    fn is_ip_unspecified(&self) -> bool {
+        // Note(unwrap): This stack only supports Ipv4.
+        self.network_interface
+            .borrow_mut()
+            .ipv4_addr()
+            .unwrap()
+            .is_unspecified()
+    }
 }
 
 impl<'a, 'b, DeviceT> embedded_nal::TcpStack for NetworkStack<'a, 'b, DeviceT>
@@ -119,6 +231,11 @@ where
         &self,
         _mode: embedded_nal::Mode,
     ) -> Result<smoltcp::socket::SocketHandle, NetworkError> {
+        // If we do not have a valid IP address yet, do not open the socket.
+        if self.is_ip_unspecified() {
+            return Err(NetworkError::NoIpAddress);
+        }
+
         match self.unused_handles.borrow_mut().pop() {
             Some(handle) => {
                 // Abort any active connections on the handle.
@@ -137,6 +254,13 @@ where
         socket: smoltcp::socket::SocketHandle,
         remote: embedded_nal::SocketAddr,
     ) -> Result<smoltcp::socket::SocketHandle, NetworkError> {
+        // If there is no longer an IP address assigned to the interface, do not allow usage of the
+        // socket.
+        if self.is_ip_unspecified() {
+            self.close(socket)?;
+            return Err(NetworkError::NoIpAddress);
+        }
+
         let mut sockets = self.sockets.borrow_mut();
         let internal_socket: &mut smoltcp::socket::TcpSocket = &mut *sockets.get(socket);
 
@@ -158,7 +282,10 @@ where
 
                 internal_socket
                     .connect((address, remote.port()), self.get_ephemeral_port())
-                    .map_err(|_| NetworkError::ConnectionFailure)?;
+                    .or_else(|_| {
+                        self.close(socket)?;
+                        Err(NetworkError::ConnectionFailure)
+                    })?;
                 Ok(socket)
             }
 
@@ -168,6 +295,12 @@ where
     }
 
     fn is_connected(&self, socket: &smoltcp::socket::SocketHandle) -> Result<bool, NetworkError> {
+        // If there is no longer an IP address assigned to the interface, do not allow usage of the
+        // socket.
+        if self.is_ip_unspecified() {
+            return Err(NetworkError::NoIpAddress);
+        }
+
         let mut sockets = self.sockets.borrow_mut();
         let socket: &mut smoltcp::socket::TcpSocket = &mut *sockets.get(*socket);
         Ok(socket.may_send() && socket.may_recv())
@@ -178,6 +311,12 @@ where
         socket: &mut smoltcp::socket::SocketHandle,
         buffer: &[u8],
     ) -> embedded_nal::nb::Result<usize, NetworkError> {
+        // If there is no longer an IP address assigned to the interface, do not allow usage of the
+        // socket.
+        if self.is_ip_unspecified() {
+            return Err(embedded_nal::nb::Error::Other(NetworkError::NoIpAddress));
+        }
+
         let mut sockets = self.sockets.borrow_mut();
         let socket: &mut smoltcp::socket::TcpSocket = &mut *sockets.get(*socket);
         socket
@@ -190,6 +329,12 @@ where
         socket: &mut smoltcp::socket::SocketHandle,
         buffer: &mut [u8],
     ) -> embedded_nal::nb::Result<usize, NetworkError> {
+        // If there is no longer an IP address assigned to the interface, do not allow usage of the
+        // socket.
+        if self.is_ip_unspecified() {
+            return Err(embedded_nal::nb::Error::Other(NetworkError::NoIpAddress));
+        }
+
         let mut sockets = self.sockets.borrow_mut();
         let socket: &mut smoltcp::socket::TcpSocket = &mut *sockets.get(*socket);
         socket