+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 This document describes the changes to smoltcp-nal between releases.
 
 # Unreleased
+* Added a reset API to close all sockets and reset DHCP whenever a link is lost. Updated DHCP to
+  close sockets if local address changes.
 * Upgraded to 0.6.1 of heapless to address security vulnerability
+* Adding support for DHCP IP assignment and management.
 
 ## Version 0.1.0
 Version 0.1.0 was published on 2021-02-17